build: fix snk test build for 32bit system

+++ c_src/quicer_nif.c
@@ -570,14 +570,14 @@ void
 resource_conn_dealloc_callback(__unused_parm__ ErlNifEnv *env, void *obj)
 {
   QuicerConnCTX *c_ctx = (QuicerConnCTX *)obj;
-  TP_CB_3(start, c_ctx->Connection, 0);
+  TP_CB_3(start, (uintptr_t)c_ctx->Connection, 0);
   AcceptorQueueDestroy(c_ctx->acceptor_queue);
   enif_free_env(c_ctx->env);
   enif_mutex_destroy(c_ctx->lock);
   CXPLAT_FREE(c_ctx->TlsSecrets, QUICER_TLS_SECRETS);
   CXPLAT_FREE(c_ctx->ssl_keylogfile, QUICER_TRACE);
   AcceptorDestroy(c_ctx->owner);
-  TP_CB_3(end, c_ctx->Connection, 0);
+  TP_CB_3(end, (uintptr_t)c_ctx->Connection, 0);
 }
 
 void
@@ -593,10 +593,10 @@ resource_conn_down_callback(__unused_parm__ ErlNifEnv *env,
     }
   else
     {
-      TP_CB_3(start, c_ctx->Connection, (uint64_t)ctx);
+      TP_CB_3(start, (uintptr_t)c_ctx->Connection, (uintptr_t)ctx);
       MsQuic->ConnectionShutdown(
           c_ctx->Connection, QUIC_CONNECTION_SHUTDOWN_FLAG_NONE, 0);
-      TP_CB_3(end, c_ctx->Connection, (uint64_t)ctx);
+      TP_CB_3(end, (uintptr_t)c_ctx->Connection, (uintptr_t)ctx);
     }
 }
 
@@ -604,12 +604,12 @@ void
 resource_stream_dealloc_callback(__unused_parm__ ErlNifEnv *env, void *obj)
 {
   QuicerStreamCTX *s_ctx = (QuicerStreamCTX *)obj;
-  TP_CB_3(start, s_ctx->Stream, 0);
+  TP_CB_3(start, (uintptr_t)s_ctx->Stream, 0);
   enif_mutex_lock(s_ctx->lock);
   enif_free_env(s_ctx->env);
   enif_mutex_unlock(s_ctx->lock);
   enif_mutex_destroy(s_ctx->lock);
-  TP_CB_3(end, s_ctx->Stream, 0);
+  TP_CB_3(end, (uintptr_t)s_ctx->Stream, 0);
 }
 
 void
@@ -626,7 +626,7 @@ resource_stream_down_callback(__unused_parm__ ErlNifEnv *env,
       return;
     }
 
-  TP_CB_3(start, s_ctx->Stream, 0);
+  TP_CB_3(start, (uintptr_t)s_ctx->Stream, 0);
   if (QUIC_FAILED(status = MsQuic->StreamShutdown(
                       s_ctx->Stream,
                       QUIC_STREAM_SHUTDOWN_FLAG_IMMEDIATE
@@ -634,11 +634,11 @@ resource_stream_down_callback(__unused_parm__ ErlNifEnv *env,
                           | QUIC_STREAM_SHUTDOWN_FLAG_ABORT_SEND,
                       0)))
     {
-      TP_CB_3(shutdown_fail, s_ctx->Stream, status);
+      TP_CB_3(shutdown_fail, (uintptr_t)s_ctx->Stream, status);
     }
   else
     {
-      TP_CB_3(shutdown_success, s_ctx->Stream, status);
+      TP_CB_3(shutdown_success, (uintptr_t)s_ctx->Stream, status);
     }
 }
 
@@ -946,7 +946,7 @@ connection_controlling_process(ErlNifEnv *env,
                                const ErlNifPid *caller,
                                const ERL_NIF_TERM *pid)
 {
-  TP_NIF_3(enter, c_ctx->Connection, (uint64_t)&c_ctx);
+  TP_NIF_3(enter, (uintptr_t)c_ctx->Connection, (uintptr_t)&c_ctx);
   if (0 != enif_compare_pids(&c_ctx->owner->Pid, caller))
     {
       return ERROR_TUPLE_2(ATOM_NOT_OWNER);
@@ -967,7 +967,7 @@ connection_controlling_process(ErlNifEnv *env,
       return ERROR_TUPLE_2(ATOM_OWNER_DEAD);
     }
 
-  TP_NIF_3(exit, c_ctx->Connection, (uint64_t)&c_ctx);
+  TP_NIF_3(exit, (uintptr_t)c_ctx->Connection, (uintptr_t)&c_ctx);
   return ATOM_OK;
 }
 
@@ -978,7 +978,7 @@ stream_controlling_process(ErlNifEnv *env,
                            const ERL_NIF_TERM *pid)
 {
 
-  TP_NIF_3(enter, s_ctx->Stream, (uint64_t)&s_ctx->owner->Pid);
+  TP_NIF_3(enter, (uintptr_t)s_ctx->Stream, (uintptr_t)&s_ctx->owner->Pid);
   if (0 != enif_compare_pids(&s_ctx->owner->Pid, caller))
     {
       return ERROR_TUPLE_2(ATOM_NOT_OWNER);
@@ -999,7 +999,7 @@ stream_controlling_process(ErlNifEnv *env,
       return ERROR_TUPLE_2(ATOM_OWNER_DEAD);
     }
 
-  TP_NIF_3(exit, s_ctx->Stream, (uint64_t)&s_ctx->owner->Pid);
+  TP_NIF_3(exit, (uintptr_t)s_ctx->Stream, (uintptr_t)&s_ctx->owner->Pid);
   return ATOM_OK;
 }
